demo3_llm_server/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.1
llama-cpp-python==0.3.8
